MDEV-25488 thd::transaction::modified_non_trans_tables is not set before ..
.. Rows_log_event caching.

In ROW binlog format a successful record modification in Engine
did not follow immediately to raising
  THD::transaction::stmt::modified_non_trans_tables (MNTT)
flag which later led to an assert in MYSQL_BIN_LOG::write_cache().
The assert also required an error at logging Rows event into
the non-transactional cache.

To avoid the assert when a non-transactional engine table is being
modified the MNTT flag must be made UP at once after the engine
changed a record. Having it UP an error in writing to the cache
is recorded to be converted into an Incident event at write_cache()
time.